:art: moves bento layout into its own component
