+++ components/dashboard/AdsCarousel.tsx
